Complete DiveCalc.calculate(min1, depth1, min2, depth2) --- untested
DiveCalc.js is to replace index.js. Right now DiveCalc is completely untested
and it is not hooked up to the gui. DiveCalc is a client side class which
computes diving information. Right now it has one possible computation:
compute the minimum wait time between dives given the minutes and depths of the
dives.

This class is untested thusfar and has work to be done on it yet to make it
more useful (todo file upcomming).